+++ Sources/Support.swift
@@ -34,6 +34,7 @@ enum SettingsKeys {
     static let faAmd = "faAmd"
     static let persistCache = "persistCache"
     static let multiGPU = "multiGPU"
+    static let forcePrivateBuffers = "forcePrivateBuffers"
     static let cacheReuse = "cacheReuse"
     static let apiKeyEnabled = "apiKeyEnabled"
     static let localNetworkDiscovery = "localNetworkDiscovery"